move all providers into compositeprovider
