+++ providers/index.tsx
@@ -1,14 +1,15 @@
 'use client'
-import { useContext, useRef, ReactNode, createContext, useState, useCallback } from 'react'
-import { LayoutRouterContext } from 'next/dist/shared/lib/app-router-context'
-import { ThemeProvider } from "next-themes"
-import { useEffect } from "react";
+import { useEffect, ReactNode } from "react";
 import utilities from 'lib/util';
-import { AnimatePresence, motion, useAnimationFrame } from 'framer-motion';
+import { AnimatePresence, motion } from 'framer-motion';
 import { usePathname } from 'next/navigation';
-import usePrefersReducedMotion from '@/hooks/usePreferesReducedMotion';
-import PaletteContextProvider, { usePalette } from 'providers/palette-context';
-import palettes from '@/styles/palettes';
+// providers
+import { ThemeProvider } from "next-themes"
+import PaletteContextProvider from './palette-context';
+import AudioContextProvider from './audio-context';
+import FrozenRouterContextProvider from './frozen-router-context';
+// providers
+
 const { setCustomProperties, toCartesianCoords } = utilities
 
 const handleInput = (event: MouseEvent | TouchEvent) => {
@@ -33,165 +34,22 @@ const handleInput = (event: MouseEvent | TouchEvent) => {
 }
 
 
-function FrozenRouter({ children }: { children: ReactNode }) {
-  const context = useContext(LayoutRouterContext)
-  const frozen = useRef(context).current;
-  return (
-    <LayoutRouterContext.Provider value={frozen}>
-      {children}
-    </LayoutRouterContext.Provider>
-  )
-}
-
-interface AudioContextType {
-  playing: boolean;
-  setPlaying: React.Dispatch<React.SetStateAction<boolean>>;
-  audioData: React.MutableRefObject<AnalyserNode | null>;
-  startPlaying: () => void;
-  stopPlaying: () => void;
-}
-
-const AudioContext = createContext<AudioContextType | null>(null);
-
-export function useAudioContext() {
-  const context = useContext(AudioContext);
-  if (!context) {
-    // This error will be thrown if the component is not a child of AudioProvider.
-    throw new Error('useAudio must be used within an AudioProvider');
-  }
-  return context;
-}
-
-function AudioProvider({ children }: { children: ReactNode }) {
-  const audioElement = useRef<HTMLAudioElement>(null);
-  const audioData = useRef<AnalyserNode | null>(null);
-  const [playing, setPlaying] = useState(false)
-  const audioContext = useRef<AudioContext | null>(null);
-  const { palette } = usePalette()
-  const musicType = palettes[palette].audio
-
-  const startPlaying = () => {
-    if (playing || !musicType) return
-    // if there is no audio element or audio context, create them
-    if (!audioContext.current) {
-      createAudioContext()
-    }
-    // audioData is created in createAudioContext
-    if (!audioElement.current || !audioData.current) return
-    // set the source of the audio element to the music type
-    audioElement.current.src = musicType
-    // 
-    audioElement.current.play()
-    setPlaying(true)
-  }
-
-  const audioAnimation = useCallback((audioData) => {
-    // create a new array of 32 bit floating point numbers
-    if (!audioData.current) return
-    let data = new Float32Array(audioData.current.frequencyBinCount)
-    // draw the audio data
-    const draw = (data: number) => {
-      const val = 0.33 + data / 3
-      utilities.setCustomProperties({
-        '--amp-primary': String(val),
-        '--amp-secondary': String(val),
-        '--amp-tertiary': String(val),
-      })
-    }
-    if (!audioData.current) return
-    audioData.current.getFloatTimeDomainData(data)
-    let sumQuares = 0.0
-    for (const ampliltude of data) {
-      sumQuares += ampliltude * ampliltude
-    }
-    const amp = Math.sqrt(sumQuares / data.length)
-    draw(amp)
-  }, [])
-
-  const noAudioAnimation = useCallback((time: number) => {
-    const now = time / 1000
-
-    // Normalize sin and cosine values to [0, 1] range
-    const primaryAmp = (Math.sin(2 * Math.PI * now / 15) + 1) / 2;
-    const secondaryAmp = (Math.sin(2 * Math.PI * now / 30) + 1) / 2;
-    const tertiaryAmp = (Math.sin(2 * Math.PI * now / 60) + 1) / 2;
-
-    // Scale and translate to desired range, for example, [1, 1.5]
-    const primaryValue = primaryAmp / 2;
-    const secondaryValue = secondaryAmp / 2;
-    const tertiaryValue = tertiaryAmp / 2;
-    utilities.setCustomProperties({
-      '--amp-primary': String(primaryValue),
-      '--amp-secondary': String(secondaryValue),
-      '--amp-tertiary': String(tertiaryValue),
-    })
-  }, [])
-
-  const stopPlaying = useCallback(() => {
-    if (!audioElement.current) return
-    audioElement.current.pause()
-    setPlaying(false)
-  }, [setPlaying])
-
-  const createAudioContext = () => {
-    audioContext.current = new window.AudioContext();
-    if (!audioElement.current) return
-    const source = audioContext.current.createMediaElementSource(audioElement.current)
-    const analyzer = audioContext.current.createAnalyser()
-    analyzer.fftSize = 2048
-    source.connect(audioContext.current.destination)
-    source.connect(analyzer)
-    audioData.current = analyzer
-  }
-  const value = {
-    playing,
-    setPlaying,
-    audioData,
-    startPlaying,
-    stopPlaying,
-  }
-
-  const shouldAnimate = !usePrefersReducedMotion()
-  useAnimationFrame((time) => {
-    if (!shouldAnimate) return
-    if (playing) {
-      audioAnimation(audioData)
-    } else {
-      noAudioAnimation(time)
-    }
-  })
-  return (
-    <AudioContext.Provider value={value}>
-      {children}
-      <audio id="audio" ref={audioElement}></audio>
-    </AudioContext.Provider>
-  )
-
-}
 
 
-const CompositeProvider = ({ children, pathname }: { children: ReactNode, pathname: string }) => {
-  return (
-    <AnimatePresence>
-      <motion.div
-        key={pathname}
-        initial={{ opacity: 0 }}
-        animate={{ opacity: 1 }}
-        exit={{ opacity: 0 }}
-        transition={{ duration: 0.4, type: 'tween' }}
-      >
-        <FrozenRouter>
-          <ThemeProvider>
-            <PaletteContextProvider>
-              <AudioProvider>
-                {children}
-              </AudioProvider>
-            </PaletteContextProvider>
-          </ThemeProvider>
-        </FrozenRouter>
-      </motion.div>
-    </AnimatePresence>
-  )
+const providers = [
+  { provider: FrozenRouterContextProvider, props: {} },
+  { provider: ThemeProvider, props: {} },
+  { provider: PaletteContextProvider, props: {} },
+  { provider: AudioContextProvider, props: {} },
+]
+
+const CompositeProvider = ({ children }: { children: ReactNode }) => {
+  return providers.reduceRight(
+    (kids, { provider: CurrentProvider, props }) => (
+      <CurrentProvider {...props}>{kids}</CurrentProvider>
+    ),
+    children
+  );
 }
 
 
@@ -211,9 +69,19 @@ export function Providers({ children }) {
   }, []);
 
   return (
-    <CompositeProvider pathname={pathname}>
-      {children}
-    </CompositeProvider>
+    <AnimatePresence>
+      <motion.div
+        key={pathname}
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+        exit={{ opacity: 0 }}
+        transition={{ duration: 0.4, type: 'tween' }}
+      >
+        <CompositeProvider>
+          {children}
+        </CompositeProvider>
+      </motion.div>
+    </AnimatePresence>
   )
 }
